♻️ refactor[#13] : 리뷰 Repository 추상화
