+++ src/main/java/com/fouribnb/review/infrastructure/jpa/ReviewJpaRepository.java
@@ -1,17 +1,13 @@
-package com.fouribnb.review.domain.repository;
+package com.fouribnb.review.infrastructure.jpa;
 
 import com.fouribnb.review.domain.entity.Review;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository {
-
-    Review save(Review review);
+public interface ReviewJpaRepository extends JpaRepository<Review, UUID> {
 
     List<Review> getAllByLodgeId(UUID lodgeId);
 
     List<Review> getAllByUserId(Long userId);
-
-    Optional<Review> findById(UUID reviewId);
 }
